refactor: JS 컴포넌트 네이밍 개선 - kebab-case로 통일 - main.js → app.js - eventHandler.js → event-handler.js - renderKanban.js → kanban-renderer.js - renderPanel.js → panel-renderer.js - template.js → template-utils.js
