+++ index.html
@@ -7,7 +7,7 @@
     <title>TASKIFY</title>
     <link rel="stylesheet" href="./styles/style.css">
 
-    <script type="module" src="./js/main.js" defer></script>
+    <script type="module" src="./js/app.js" defer></script>
 </head>
 
 <body class="flex-column">
