--- js/main.js
@@ -1,7 +1,0 @@
-import { initKanban } from './components/renderKanban.js'
-import { initPanel } from './components/renderPanel.js'
-import { initEventHandlers } from './components/eventHandler.js';
-
-initPanel();
-initKanban();
-initEventHandlers();
